+++ src/Components/index.ts
@@ -1,2 +1,3 @@
 export {default as NavLinks} from "./NavLinks/NavLinks";
 export {default as AboutMe} from "./AboutMe/AboutMe";
+export {default as Skills} from "./Skills/Skills";